Update stager-newrelic-java
Removed the apcera namespace from examples to prevent conflicts
in a shared environment.

Removed code specific to dealing with unintentionally nesting
directories (from the java stager putting the files into an /app
directory and this stager repeating it).  Can remove this because
version 0.3.0 of our stager should not be automatically adding the
extra file depth.